silent phones

+++ FastSpeech2/phonetise/phonetise_arabic.py
@@ -376,7 +376,7 @@ def phonetise(text):
 					iterations = 1
 					for index, letter in enumerate(phones):
 						if(isinstance(letter, list)):
-							curIndex = (i / iterations) % len(letter)
+							curIndex = int((i / iterations) % len(letter))
 							if(letter[curIndex] != u''):
 								pronunciations[-1].append(letter[curIndex])
 							iterations = iterations * len(letter)
